requirements.txt
accelerate
bitsandbytes
diffusers>=0.32.1
transformers>=4.45.2
huggingface_hub
hf_transfer>=0.1.8
peft>=0.13.0
decord>=0.6.0
wandb
pandas
torchao>=0.7.0
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
